Setup static roles pr fixes
